requirements.txt
aiofiles==24.1.0
blinker==1.8.2
click==8.1.7
dnspython==2.7.0
Flask==3.0.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
Hypercorn==0.17.3
hyperframe==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
nats-py==2.9.0
priority==2.0.0
pymongo==4.10.1
python-dotenv==1.0.1
Quart==0.19.6
quart-cors==0.7.0
quart-uploads==0.0.4
types-aiofiles==24.1.0.20240626
Werkzeug==3.0.4
wsproto==1.2.0
